+++ GospoRol.Application/ViewModels/FieldViews/NewFieldVm.cs
@@ -14,13 +14,13 @@ namespace GospoRol.Application.ViewModels
         [DisplayName("Nazwa Pola")]
         public string FieldName { get; set; }
         [DisplayName("Areał")]
-        public double Acreage { get; set; }
+        public decimal Acreage { get; set; }
         [DisplayName("Uprawiana roślina")]
         public string CultivatedPlant { get; set; }
         [DisplayName("Odmiana uprawianej rośliny")]
         public string Variety { get; set; }    
         [DisplayName("Odległość do magazynu")]
-        public double DistanceToWarehouse { get; set; }
+        public decimal DistanceToWarehouse { get; set; }
         [DisplayName("Klasa gleby")]
         public int AgriculturalClassId { get; set; }
         public List<SelectListItem> AgriculturalClasses { get; set; }
